requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
flask-sqlalchemy==3.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
flask-login==0.6.2
requests==2.31.0
python-dateutil==2.8.2
flask-migrate==4.0.4
flask-wtf==1.1.1
gunicorn==21.2.0
